% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choisir_difficulte.R
\name{choisir_difficulte}
\alias{choisir_difficulte}
\title{Choisir le nombre de cases préremplies en fonction de la difficulté}
\usage{
choisir_difficulte(taille, niveau)
}
\arguments{
\item{taille}{Un entier représentant la taille de la grille (doit être un multiple de 2 entre 4 et 100).}

\item{niveau}{Un caractère indiquant le niveau de difficulté. Doit être l'un des suivants :
- `"Facile"` (50% des cases remplies)
- `"Moyen"` (40% des cases remplies)
- `"Difficile"` (30% des cases remplies)
- `"Einstein"` (10% des cases remplies, très difficile)}
}
\value{
Un entier correspondant au nombre de cases à préremplir.
}
\description{
Cette fonction détermine combien de cases doivent être préremplies dans une
grille de Takuzu en fonction de la taille et du niveau de difficulté choisi.
}
