% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Lancer l'application Shiny du package Takuzu}
\usage{
run_app()
}
\value{
Aucune valeur de retour. Cette fonction a pour effet de démarrer une session Shiny.
}
\description{
Cette fonction permet de lancer l'application Shiny embarquée dans le package.
L'exécute via `shiny::runApp()`. Elle est destinée à être utilisée par l'utilisateur
final du package.
}
\examples{
if (interactive()) {
  run_app()
}

}
