% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_takuzu.R
\name{generer_takuzu}
\alias{generer_takuzu}
\title{Génère une grille Takuzu valide avec un niveau de difficulté donné}
\usage{
generer_takuzu(taille, niveau)
}
\arguments{
\item{taille}{Un entier pair représentant la taille de la grille (ex: 6, 8, 10).}

\item{niveau}{Un caractère indiquant le niveau de difficulté : "facile", "moyen", ou "difficile".}
}
\value{
Une matrice de taille x taille contenant la grille Takuzu avec certaines cases vides.
}
\description{
Génère une grille Takuzu valide avec un niveau de difficulté donné
}
